+++ node/ts/NiceConverters.ts
@@ -6,6 +6,8 @@
 import { ServiceId } from './Address.js';
 import { type CdnCredentials } from './net/chat/CdnCredentials.js';
 import * as Native from './Native.js';
+import { newNativeHandle, wrapStream } from './internal.js';
+import { type TokioAsyncContext } from './net.js';
 
 export type DeviceId = number;
 /**
@@ -28,6 +30,19 @@ export function cdnCredentialReturnConverter(
 ): CdnCredentials {
   return {
     headers: new Map(headers),
+  };
+}
+
+export function copyBackupMediaStreamConverter(
+  streamHandle: Native.CopyBackupMediaStream
+): (
+  asyncContext: TokioAsyncContext
+) => ReadableStream<Native.ReturnFfiBridgeCopyBackupMediaOutcome> {
+  return (asyncContext) => {
+    return wrapStream(asyncContext, newNativeHandle(streamHandle), {
+      pull: Native.CopyBackupMediaStream_next,
+      cancel: Native.CopyBackupMediaStream_cancel,
+    });
   };
 }