+++ node/ts/test/chat/UnauthBackupsServiceTest.ts
@@ -7,9 +7,14 @@ import { config, expect, use } from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 
 import * as Native from '../../Native.js';
+import * as NativeNice from '../../NativeNice.js';
 import * as util from '../util.js';
 import { TokioAsyncContext, UnauthBackupsService } from '../../net.js';
-import { connectUnauth, testSimpleGrpcRequest } from './ServiceTestUtils.js';
+import {
+  connectUnauth,
+  defineTestGrpcCases,
+  testSimpleGrpcRequest,
+} from './ServiceTestUtils.js';
 import {
   BackupAuthCredential,
   GenericServerPublicParams,
@@ -321,6 +326,56 @@ describe('UnauthBackupsService', () => {
           auth: TEST_AUTH,
           rng: { __deterministicRngSeedForTesting: 0 },
         })
+    );
+  });
+
+  describe('copyBackupMedia', () => {
+    defineTestGrpcCases(
+      NativeNice.TESTING_CopyBackupMediaTests(),
+      connectUnauth<UnauthBackupsService>,
+      async (chat, items, expected) => {
+        Native.TESTING_EnableDeterministicRngForTesting();
+
+        const stream = chat.copyBackupMedia({
+          auth: TEST_AUTH,
+          items: items.map((next) => ({
+            ...next,
+            objectLength: Number(next.objectLength),
+          })),
+          rng: { __deterministicRngSeedForTesting: 0 },
+        });
+        const [actualItems, maybeError] = await util.collectUntilError(stream);
+
+        for (const nextExpected of expected) {
+          if (typeof nextExpected === 'object') {
+            const actualItem = actualItems.shift();
+            expect(actualItem?.mediaId).deep.equals(nextExpected.item.mediaId);
+            const expectedResult = nextExpected.item.result;
+            expect(actualItem?.result).deep.equals(
+              typeof expectedResult === 'object'
+                ? { cdn: expectedResult.success }
+                : expectedResult
+            );
+          } else {
+            switch (nextExpected) {
+              case 'invalidDataInStream':
+              case 'credentialRejectedWithoutAppropriateServerInfo':
+                expect(maybeError)
+                  .instanceOf(LibSignalErrorBase)
+                  .with.property('code', ErrorCode.IoError);
+                break;
+              case 'credentialRejected':
+                expect(maybeError)
+                  .instanceOf(LibSignalErrorBase)
+                  .with.property('code', ErrorCode.RequestUnauthorized);
+                break;
+              default:
+                nextExpected satisfies never;
+            }
+          }
+        }
+        expect(actualItems).deep.equals([]);
+      }
     );
   });
 });