+++ rust/bridge/shared/types/src/node/convert.rs
@@ -2359,6 +2359,25 @@ where
     }
 }
 
+#[cfg(feature = "metadata")]
+impl NiceResultConverter for crate::net::chat::CopyBackupMediaStream {
+    fn register_ts_result_converter(_ctx: &mut TsMetadataContext) -> TsReturnConverter {
+        TsReturnConverter {
+            nice_type: concat!(
+                "(",
+                "(asyncContext: TokioAsyncContext)",
+                // Note that we delay converting this to a nice type for TS module dependency cycle
+                // reasons.
+                "=> ReadableStream<Native.ReturnFfiBridgeCopyBackupMediaOutcome>",
+                ")"
+            )
+            .to_owned(),
+            ffi_type: "Native.CopyBackupMediaStream".to_owned(),
+            converter_function: "copyBackupMediaStreamConverter".to_owned(),
+        }
+    }
+}
+
 /// The name of the property on JavaScript objects that wrap a boxed Rust value.
 pub(crate) const NATIVE_HANDLE_PROPERTY: &str = "_nativeHandle";