+++ RELEASE_NOTES.md
@@ -1,4 +1,5 @@
 v0.97.4
 
 - AuthAccountsService.setRegistrationLock() typed API
+- UnauthBackupsService.copyBackupMedia() typed API
 - node: Add full-er service SVR2 APIs that performs necessary crypto operations inside libsignal.